+++ js/controls.js
@@ -387,6 +387,33 @@ class Controls {
         });
     }
     /**
+     * Deduce control type based on the params
+     */
+    deduceControlType(params) {
+        const value = params.value;
+        if (typeof value === 'boolean') {
+            return 'boolean';
+        }
+        if (typeof value === 'number' ||
+            // Allow numeric strings with units limited to px, em, rem, %
+            (typeof value === 'string' &&
+                /^-?\d+\.?\d*\s*(px|em|rem|%)$/.test(value))) {
+            return 'number';
+        }
+        if (Array.isArray(params.options)) {
+            return 'select';
+        }
+        if (typeof value === 'string') {
+            if (params.path.toLowerCase().indexOf('color') !== -1 ||
+                Product.color(value).rgba.toString().indexOf('NaN') === -1) {
+                return 'color';
+            }
+            return 'text';
+        }
+        // Default to text
+        return 'text';
+    }
+    /**
      * Add a control
      */
     addControl(params) {
@@ -398,6 +425,7 @@ class Controls {
             const targetOptions = this.target.getOptions?.();
             params.value = (targetOptions && getNestedValue(targetOptions, params.path)) ?? getNestedValue(Product?.defaultOptions, params.path);
         }
+        params.type || (params.type = this.deduceControlType(params));
         const div = this.container.appendChild(Object.assign(document.createElement('div'), { className: 'hcc-control' }));
         const keyDiv = div.appendChild(Object.assign(document.createElement('div'), { className: 'hcc-key' }));
         const valueDiv = div.appendChild(Object.assign(document.createElement('div'), { className: 'hcc-value' }));
@@ -495,7 +523,7 @@ class HighchartsControlsElement extends HTMLElement {
         const controls = [];
         this.querySelectorAll('highcharts-control').forEach((controlEl) => {
             const control = controlEl.getConfig();
-            if (control.type && control.path) {
+            if (control.path) {
                 controls.push(control);
             }
         });